CLI Options added, modified IO to take absolute out paths

+++ src/uk/co/ohpollux/space/program/IO.java
@@ -1,9 +1,7 @@
 package uk.co.ohpollux.space.program;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.PrintWriter;
 
 import org.apache.commons.io.FileUtils;
 
@@ -22,27 +20,22 @@ public class IO {
 	return result;
     }
 
+    @SuppressWarnings("deprecation")
     public static boolean saveStringAsFile(String text, String path) {
 	boolean success = false;
-	PrintWriter out = null;
-	try {
+
+	if (path != null) {
 	    File file = new File(path);
 
-	    if (!file.exists()) {
-		success = file.mkdirs();
-	    } else {
+	    if (file.exists()) {
 		file.delete();
-		file.mkdirs();
 	    }
 
-	    out = new PrintWriter(path);
-	    out.println(text);
-
-	} catch (FileNotFoundException e) {
-	    success = false;
-	} finally {
-	    if (out != null) {
-		out.close();
+	    try {
+		FileUtils.writeStringToFile(file, text);
+		success = true;
+	    } catch (IOException e) {
+		success = false;
 	    }
 	}
